+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - **! Added `Table.contains(self: Self@Table, attribute_name: str, attribute_value: Any) -> bool`.**
 - **! Added `Table.not_contains(self: Self@Table, attribute_name: str, attribute_value: Any) -> bool`.**
+- Added `Table.add_rows(self: Self@Table, rows: Iterable[tuple[Any, ...]], *, lock: bool = True) -> None`.
 - Added `Table.get_attribute_index(self: Self@Table, name: str) -> int`.
 
 ## [0.2.0] - 2022-11-19